modify notes
